refactor: move Console Repositories to repository package
